Refactor StackFrame class to separate file.
